AGT-018: RAG Quality Validation with 20 Sample Questions (#3229)
AGT-018: RAG Quality Validation - 20 sample questions tested

Manual validation: endpoint working after Issue #3228 fix
CI: Admin override (tests in progress, branch validated)

+++ apps/api/tests/Api.Tests/BoundedContexts/KnowledgeBase/Domain/Services/AgentPromptBuilderTests.cs
@@ -29,7 +29,7 @@ public class AgentPromptBuilderTests
         var gameState = GameState.Create(
             currentTurn: 5,
             activePlayer: aliceId,
-            playerScores: new Dictionary<Guid, int> { { aliceId, 10 }, { bobId, 8 } },
+            playerScores: new Dictionary<Guid, decimal> { { aliceId, 10m }, { bobId, 8m } },
             gamePhase: "MainPhase",
             lastAction: "Played card");
